all singletons dbs, currency product changes

+++ app/src/main/java/com/example/myapplication/MainViewModel.kt
@@ -1,51 +1,49 @@
 package com.example.myapplication
 
-import android.app.Application
-import androidx.lifecycle.AndroidViewModel
+import android.util.Log
+import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.myapplication.profile.currency.currencyLocal.CurrencyDatabase
 import com.example.myapplication.profile.currency.currencyLocal.CurrencyLocalModel
 import com.example.myapplication.profile.currency.currencyLocal.CurrencyLocalRepository
 import com.example.myapplication.profile.currency.currencyLocal.MapTypeConverter
 import com.example.myapplication.profile.currency.currencyRetrofit.CurrencyModel
 import com.example.myapplication.profile.currency.currencyRetrofit.CurrencyRepository
+import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import javax.inject.Inject
 
-class MainViewModel(private val repository: CurrencyRepository, application: Application) :
-    AndroidViewModel(application) {
-
-    private val localRepo: CurrencyLocalRepository
-
-    init {
-        val currencyDao = CurrencyDatabase.getDatabase(application).currencyDao()
-
-        localRepo = CurrencyLocalRepository(currencyDao)
-
-        getRates()
-    }
+@HiltViewModel
+class MainViewModel@Inject constructor(private val localRepo: CurrencyLocalRepository, private val retrofit: CurrencyRepository) :ViewModel()  {
 
     private fun insertCurrency(currencyModel: CurrencyModel) {
         viewModelScope.launch(Dispatchers.IO) {
 
-            localRepo.deleteDataBase()
-            var mapTypeConverter = MapTypeConverter()
-            localRepo.insertDataBase(
-                CurrencyLocalModel(
-                    currencyModel.success,
-                    currencyModel.timestamp,
-                    currencyModel.base,
-                    currencyModel.date,
-                    mapTypeConverter.RatestoMap(currencyModel.rates)!!
+            if (localRepo.readAllData()[0].date != currencyModel.date) {
+                localRepo.deleteDataBase()
+                val mapTypeConverter = MapTypeConverter()
+                localRepo.insertDataBase(
+                    CurrencyLocalModel(
+                        currencyModel.success,
+                        currencyModel.timestamp,
+                        currencyModel.base,
+                        currencyModel.date,
+                        mapTypeConverter.RatestoMap(currencyModel.rates)!!
+                    )
                 )
-            )
+                Log.d("delul", localRepo.readAllData()[0].date)
+                Log.d("delulr", currencyModel.date)
 
+            } else {
+                Log.d("lul", localRepo.readAllData()[0].date)
+                Log.d("lulr", currencyModel.date)
+            }
         }
     }
 
-    private fun getRates() {
+    fun getRates() {
         viewModelScope.launch {
-            val response = repository.getRates()
+            val response = retrofit.getRates()
             insertCurrency(response)
         }
     }
